+++ src/main/java/com/example/codionbe/domain/member/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.example.codionbe.domain;
+package com.example.codionbe.domain.member.repository;
 
-import com.example.codionbe.domain.member.User;
+import com.example.codionbe.domain.member.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
